requirements.txt
asgiref==3.2.7
autopep8==1.5.1
coverage==5.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.8
django-adminlte2==0.4.1
django-stdimage==5.0.3
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
mccabe==0.6.1
model-mommy==2.0.0
Pillow==7.1.1
progressbar2==3.50.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
python-utils==2.4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
yapf==0.29.0
